# openapi.json
{"openapi":"3.1.0","info":{"title":"Legal Agent API","version":"0.1.0"},"paths":{"/copilot":{"post":{"summary":"Copilot Endpoint","operationId":"copilot_endpoint_copilot_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CopilotRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/registry/upsert":{"post":{"summary":"Registry Upsert","description":"Embed a single department registry record and upsert it into Qdrant.\nUses BGEEmbedder (falls back to Nomic if sentence-transformers unavailable).","operationId":"registry_upsert_registry_upsert_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeptRegistryUpsertRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/registry/{uuid}":{"delete":{"summary":"Registry Delete","description":"Delete a single department registry point from Qdrant by UUID.","operationId":"registry_delete_registry__uuid__delete","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string","title":"Uuid"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CopilotRequest":{"properties":{"query":{"type":"string","title":"Query"}},"type":"object","required":["query"],"title":"CopilotRequest"},"DeptRegistryUpsertRequest":{"properties":{"entity_id":{"type":"string","title":"Entity Id"},"entity_name":{"type":"string","title":"Entity Name"},"category":{"type":"string","title":"Category"},"what_they_can_provide":{"items":{"type":"string"},"type":"array","title":"What They Can Provide","default":[]},"legal_basis_typically_cited":{"items":{"type":"string"},"type":"array","title":"Legal Basis Typically Cited","default":[]},"request_format_expected":{"type":"string","title":"Request Format Expected","default":""},"typical_response_time":{"type":"string","title":"Typical Response Time","default":""},"escalation_path_if_no_response":{"type":"string","title":"Escalation Path If No Response","default":""},"notes_or_caveats":{"type":"string","title":"Notes Or Caveats","default":""},"confidence":{"type":"string","title":"Confidence","default":"high"},"act":{"type":"string","title":"Act","default":"department_registry"},"qdrant_uuid":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Qdrant Uuid"}},"type":"object","required":["entity_id","entity_name","category"],"title":"DeptRegistryUpsertRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}